refactor(accesscore): split cell_providers.go from cell_routes.go (F4 / A5a-R5)
accesscore's cell_routes.go previously mixed three distinct concerns:
- exposed service providers (HealthCheckers / TokenVerifier / Authorizer)
  consumed by runtime/auth middleware + bootstrap health probes
- HTTP routing wiring (RegisterRoutes)
- outbox event subscriptions (RegisterSubscriptions)

Split by concern, mirroring the cell.go physical split introduced in PR-A5a:
- cell_providers.go (new): HealthCheckers, TokenVerifier, Authorizer
- cell_routes.go: RegisterRoutes + RegisterSubscriptions
- cell_init.go + cell.go unchanged

configcore's cell_routes.go already contains only RegisterRoutes +
RegisterSubscriptions (no provider methods); no split needed. auditcore
likewise has no cross-cutting providers to separate.

Closes A5a-R5 CELL-ROUTES-PROVIDERS-SPLIT-01 (F4 of PR-A5c).
